backend/requirements.txt
fastapi
uvicorn
python-multipart
groq
tavily-python
python-dotenv